requirements.txt
streamlit
google-generativeai
pandas
numpy
yfinance
plotly
python-dotenv
requests
